+++ src/tests/ast/msg_string.rs
@@ -66,7 +66,6 @@ fn typed() {
     assert_eq!(
         message,
         Message {
-            resource: "test".to_string(),
             id: Id::new_msg("greeting"),
             comment: vec!["$name (String) - The name.".to_string()],
             variables: vec![Variable {