+++ src/tests/ast/msg_text.rs
@@ -57,7 +57,6 @@ fn typed() {
     assert_eq!(
         message,
         Message {
-            resource: "test".to_string(),
             id: Id::new_msg("hello-world"),
             comment: vec![],
             variables: vec![],